+++ src/main/java/com/tuioe/blog/dto/CommentDTO.java
@@ -2,7 +2,7 @@ package com.tuioe.blog.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.tuioe.blog.Entity.Comment;
+import com.tuioe.blog.domain.Entity.Comment;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -16,7 +16,7 @@ public class CommentDTO {
     private String content;
     private String nickname;
 
-    private int idx;
+    private Long idx;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime date;
